Refactor project structure by moving service files to a dedicated services directory; update import paths in controllers accordingly. Add new services for admin, appointment, auth, doctor, patient, payment, prescription, and user functionalities, enhancing modularity and maintainability of the healthcare system API.
