+++ healthcare-system/src/services/userService.js
@@ -1,5 +1,5 @@
-const User = require('../User');
-const logger = require('../../utils/logger');
+const User = require('../models/User');
+const logger = require('../utils/logger');
 
 class UserService {
     /**
